Split exercise submissions controller by context
Mirror the challenges structure: lesson create/latest now live in
Internal::Lessons::ExerciseSubmissionsController, matching
Internal::Challenges::ExerciseSubmissionsController. The naked
Internal::ExerciseSubmissionsController is left with just the
context-agnostic progression-scores PATCH (update).

Route helpers are unchanged (only the nested route's controller: option
moved), so only the handling class differs.
