+++ config/routes.rb
@@ -96,7 +96,7 @@ Rails.application.routes.draw do
     # Always have the param as lesson slug - auto-prefixed in the second
     resources :lessons, only: [:show], param: :lesson_slug
     resources :lessons, only: [], param: :slug do
-      resources :exercise_submissions, only: [:create] do
+      resources :exercise_submissions, only: [:create], controller: 'lessons/exercise_submissions' do
         collection do
           get :latest
         end
