+++ test/controllers/internal/exercise_submissions_controller_test.rb
@@ -3,253 +3,12 @@ require "test_helper"
 class Internal::ExerciseSubmissionsControllerTest < ApplicationControllerTest
   setup do
     setup_user
-    @level = create(:level)
-    @lesson = create(:lesson, :exercise, level: @level)
-    @user_level = create(:user_level, user: @current_user, level: @level)
+    @lesson = create(:lesson, :exercise)
   end
 
-  guard_incorrect_token! :internal_lesson_exercise_submissions_path, args: ["test-slug"], method: :post
-  guard_incorrect_token! :latest_internal_lesson_exercise_submissions_path, args: ["test-slug"], method: :get
   guard_incorrect_token! :internal_exercise_submission_path, args: ["test-uuid"], method: :patch
 
-  # GET /internal/lessons/:slug/exercise_submissions/latest tests
-  test "GET latest returns most recent submission" do
-    user_lesson = create(:user_lesson, user: @current_user, lesson: @lesson)
-    create(:exercise_submission, context: user_lesson)
-    latest_submission = create(:exercise_submission, context: user_lesson)
-    serialized = { uuid: "test", files: [] }
-
-    SerializeExerciseSubmission.expects(:call).with(latest_submission).returns(serialized)
-
-    get latest_internal_lesson_exercise_submissions_path(lesson_slug: @lesson.slug),
-      as: :json
-
-    assert_response :success
-    assert_json_response({ submission: serialized })
-  end
-
-  test "GET latest returns 404 when user has no user_lesson" do
-    get latest_internal_lesson_exercise_submissions_path(lesson_slug: @lesson.slug),
-      as: :json
-
-    assert_json_error(:not_found)
-  end
-
-  test "GET latest returns 404 when user has no submissions" do
-    create(:user_lesson, user: @current_user, lesson: @lesson)
-
-    get latest_internal_lesson_exercise_submissions_path(lesson_slug: @lesson.slug),
-      as: :json
-
-    assert_json_error(:not_found)
-  end
-
-  test "GET latest returns 404 for non-existent lesson" do
-    get latest_internal_lesson_exercise_submissions_path(lesson_slug: "nonexistent"),
-      as: :json
-
-    assert_json_error(:not_found, error_type: :lesson_not_found)
-  end
-
-  # POST /internal/lessons/:slug/exercise_submissions tests
-  test "POST create successfully creates submission" do
-    files = [
-      { filename: "main.rb", code: "puts 'hello'" },
-      { filename: "helper.rb", code: "def help\nend" }
-    ]
-
-    Prosopite.pause do
-      post internal_lesson_exercise_submissions_path(lesson_slug: @lesson.slug),
-        params: { submission: { files: } },
-        as: :json
-    end
-
-    assert_response :created
-    assert_json_response({ submission: { uuid: ExerciseSubmission.last.uuid } })
-  end
-
-  test "POST create starts UserLesson" do
-    files = [{ filename: "solution.rb", code: "# code" }]
-
-    UserLesson::Start.expects(:call).with(
-      @current_user,
-      @lesson
-    ).returns(create(:user_lesson))
-
-    post internal_lesson_exercise_submissions_path(lesson_slug: @lesson.slug),
-      params: { submission: { files: } },
-      as: :json
-
-    assert_response :created
-  end
-
-  test "POST create calls ExerciseSubmission::Create" do
-    user_lesson = create(:user_lesson, user: @current_user, lesson: @lesson)
-    files = [{ filename: "test.rb", code: "puts 'test'" }]
-
-    UserLesson::Start.stubs(:call).returns(user_lesson)
-
-    ExerciseSubmission::Create.expects(:call).with do |ul, file_params|
-      ul == user_lesson &&
-        file_params.length == 1 &&
-        file_params[0]["filename"] == "test.rb" &&
-        file_params[0]["code"] == "puts 'test'"
-    end.returns(create(:exercise_submission))
-
-    post internal_lesson_exercise_submissions_path(lesson_slug: @lesson.slug),
-      params: { submission: { files: } },
-      as: :json
-
-    assert_response :created
-  end
-
-  test "POST create handles invalid lesson slug" do
-    files = [{ filename: "test.rb", code: "code" }]
-
-    post internal_lesson_exercise_submissions_path(lesson_slug: "nonexistent"),
-      params: { submission: { files: } },
-      as: :json
-
-    assert_json_error(:not_found, error_type: :lesson_not_found)
-  end
-
-  test "POST create with multiple files" do
-    files = [
-      { filename: "file1.rb", code: "code1" },
-      { filename: "file2.rb", code: "code2" },
-      { filename: "file3.rb", code: "code3" }
-    ]
-
-    Prosopite.pause do
-      post internal_lesson_exercise_submissions_path(lesson_slug: @lesson.slug),
-        params: { submission: { files: } },
-        as: :json
-    end
-
-    assert_response :created
-  end
-
-  test "POST create returns 422 for duplicate filenames" do
-    files = [
-      { filename: "main.rb", code: "code1" },
-      { filename: "main.rb", code: "code2" }
-    ]
-
-    post internal_lesson_exercise_submissions_path(lesson_slug: @lesson.slug),
-      params: { submission: { files: } },
-      as: :json
-
-    assert_response :unprocessable_entity
-    assert_json_response({
-      error: {
-        type: "duplicate_filename",
-        message: "Duplicate filenames: main.rb"
-      }
-    })
-  end
-
-  test "POST create returns 422 for too many files" do
-    files = Array.new(21) { |i| { filename: "file#{i}.rb", code: "code#{i}" } }
-
-    post internal_lesson_exercise_submissions_path(lesson_slug: @lesson.slug),
-      params: { submission: { files: } },
-      as: :json
-
-    assert_response :unprocessable_entity
-    assert_json_response({
-      error: {
-        type: "too_many_files",
-        message: "Too many files (maximum 20)"
-      }
-    })
-  end
-
-  test "POST create returns 422 for file too large" do
-    files = [
-      { filename: "large.rb", code: "a" * 100_001 }
-    ]
-
-    post internal_lesson_exercise_submissions_path(lesson_slug: @lesson.slug),
-      params: { submission: { files: } },
-      as: :json
-
-    assert_response :unprocessable_entity
-
-    json_response = response.parsed_body
-    assert_equal "file_too_large", json_response["error"]["type"]
-    assert_match(/File 'large.rb' is too large/, json_response["error"]["message"])
-  end
-
-  test "POST create returns 422 for empty files array" do
-    post internal_lesson_exercise_submissions_path(lesson_slug: @lesson.slug),
-      params: { submission: { files: [] } },
-      as: :json
-
-    assert_response :unprocessable_entity
-
-    json_response = response.parsed_body
-    assert_equal "invalid_submission", json_response["error"]["type"]
-    assert_match(/at least one file/i, json_response["error"]["message"])
-  end
-
-  test "POST create returns 422 for missing filename" do
-    files = [{ code: "puts 'hello'" }]
-
-    post internal_lesson_exercise_submissions_path(lesson_slug: @lesson.slug),
-      params: { submission: { files: } },
-      as: :json
-
-    assert_response :unprocessable_entity
-
-    json_response = response.parsed_body
-    assert_equal "invalid_submission", json_response["error"]["type"]
-    assert_match(/filename.*required/i, json_response["error"]["message"])
-  end
-
-  test "POST create returns 422 for null filename" do
-    files = [{ filename: nil, code: "puts 'hello'" }]
-
-    post internal_lesson_exercise_submissions_path(lesson_slug: @lesson.slug),
-      params: { submission: { files: } },
-      as: :json
-
-    assert_response :unprocessable_entity
-
-    json_response = response.parsed_body
-    assert_equal "invalid_submission", json_response["error"]["type"]
-    assert_match(/filename.*required/i, json_response["error"]["message"])
-  end
-
-  test "POST create returns 422 for missing code" do
-    files = [{ filename: "main.rb" }]
-
-    post internal_lesson_exercise_submissions_path(lesson_slug: @lesson.slug),
-      params: { submission: { files: } },
-      as: :json
-
-    assert_response :unprocessable_entity
-
-    json_response = response.parsed_body
-    assert_equal "invalid_submission", json_response["error"]["type"]
-    assert_match(/code.*required/i, json_response["error"]["message"])
-  end
-
-  test "POST create returns 422 for null code" do
-    files = [{ filename: "main.rb", code: nil }]
-
-    post internal_lesson_exercise_submissions_path(lesson_slug: @lesson.slug),
-      params: { submission: { files: } },
-      as: :json
-
-    assert_response :unprocessable_entity
-
-    json_response = response.parsed_body
-    assert_equal "invalid_submission", json_response["error"]["type"]
-    assert_match(/code.*required/i, json_response["error"]["message"])
-  end
-
   # PATCH /internal/exercise_submissions/:uuid (progression scores) tests
-
   test "PATCH update calls UpdateProgressionScores and returns 200" do
     user_lesson = create(:user_lesson, user: @current_user, lesson: @lesson)
     submission = create(:exercise_submission, context: user_lesson)
